Migrate ax mode changes from BrowserAccessibility to AXPlatformNode.
Bug: 703369

+++ components/password_manager/core/browser/password_syncable_service.cc
@@ -516,10 +516,9 @@ bool PasswordSyncableService::ReadFromPasswordStore(
       !password_store_->FillBlacklistLogins(&blacklist_entries)) {
     // Password store often fails to load passwords. Track failures with UMA.
     // (http://crbug.com/249000)
-    // TODO(wychen): enum uma should be strongly typed. crbug.com/661401
     UMA_HISTOGRAM_ENUMERATION("Sync.LocalDataFailedToLoad",
                               ModelTypeToHistogramInt(syncer::PASSWORDS),
-                              static_cast<int>(syncer::MODEL_TYPE_COUNT));
+                              syncer::MODEL_TYPE_COUNT);
     return false;
   }
   password_entries->resize(autofillable_entries.size() +
